chore: add tdd

+++ db/migration/000001_init_whiteboard_schema.up.sql
@@ -1,6 +1,6 @@
 -- authors table
 CREATE TABLE "authors" (
-  id INT PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   name VARCHAR(255),
   email VARCHAR(255),
   password VARCHAR(255),
@@ -10,7 +10,7 @@ CREATE TABLE "authors" (
 
 -- whiteboard table
 CREATE TABLE whiteboard (
-  id INT PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   name VARCHAR(255),
   created_by INT,
   created_at timestamp not null default (now()),
@@ -20,7 +20,7 @@ CREATE TABLE whiteboard (
 
 -- stroke table
 CREATE TABLE stroke (
-  id INT PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   whiteboard_id INT,
   color VARCHAR(255),
   line_width INT,
@@ -33,7 +33,7 @@ CREATE TABLE stroke (
 
 -- point table
 CREATE TABLE point (
-  id INT PRIMARY KEY,
+  id BIGSERIAL PRIMARY KEY,
   stroke_id INT,
   x_coordinate FLOAT,
   y_coordinate FLOAT,
